agent/requirements.txt
pika
python-dotenv
openai
tiktoken
google-cloud-aiplatform
anthropic
